requirements.txt
asgiref==3.7.2
black==23.12.1
click==8.1.7
Django==4.2.11
django-filter==24.2
djangorestframework==3.15.1
django-debug-toolbar==4.2.0
drf-spectacular==0.27.2
flake8==7.0.0
isort==5.13.2
ipython==8.18.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
sqlparse==0.4.4
